+++ apps/eaas-client/.env.example
@@ -1,2 +1,4 @@
 REACT_APP_API_BASE_URL=http://localhost:4000/
 REACT_APP_PROVIDER_URL=
+REACT_APP_CHAIN_ID=
+REACT_APP_PRIVATE_KEY=
