+++ apps/eaas-client/src/App.tsx
@@ -2,18 +2,18 @@ import { ChakraProvider } from "@chakra-ui/react";
 import { BrowserRouter as Router } from "react-router-dom";
 import { AppRoutes } from "./ui";
 import { StoreProvider } from "./store/storeProvider";
-import { WalletProvider } from "./ui/embalmer/WalletProvider";
 import { theme } from "ui/theme";
+import { NetworkConfigProvider } from "ui/embalmer/NetworkConfigProvider";
 
 function App() {
   return (
     <StoreProvider>
       <ChakraProvider theme={theme}>
-        <WalletProvider>
+        <NetworkConfigProvider>
           <Router>
             <AppRoutes />
           </Router>
-        </WalletProvider>
+        </NetworkConfigProvider>
       </ChakraProvider>
     </StoreProvider>
   );
